Feat: 회원가입 로직
회원가입 시 인증 메일이 발송되도록 변경

+++ backend/src/main/java/com/pipewatch/domain/user/model/entity/EmployeeInfo.java
@@ -19,7 +19,7 @@ public class EmployeeInfo {
 	private Long id;
 
 	@NotNull
-	private Integer empNo;
+	private Long empNo;
 
 	@NotNull
 	private String department;
